fix openwebtext length

+++ timber/main/llama_eval.py
@@ -1,28 +1,21 @@
 import os
-import time
-import traceback
-import torch
-import transformers
-from datasets import load_dataset
-from tqdm import tqdm
-import argparse
-from transformers import TextStreamer
 import pathlib
 
-from vllm import LLM
-from vllm.model_executor.layers.attention import PagedAttention
+import torch
+import transformers
 from peft import LoraConfig, TaskType, PeftModel
-from peft import get_peft_model, prepare_model_for_kbit_training
-from timber.models.modeling_llama import LlamaForCausalLM, LlamaConfig
-from timber.utils import seed, get_bench
+from peft import get_peft_model
 
+from timber.main.eval_args import eval_args, ArgsType
 from timber.main.jobs.bench_single_layer import job_bench_single_layer
-from timber.main.jobs.ppl import job_ppl
-from timber.main.jobs.stream import job_stream
-from timber.main.jobs.mmlu import job_mmlu
 from timber.main.jobs.booksum import job_booksum
 from timber.main.jobs.merge_lora import job_merge_lora
-from timber.main.eval_args import eval_args, ArgsType
+from timber.main.jobs.mmlu import job_mmlu
+from timber.main.jobs.ppl import job_ppl
+from timber.main.jobs.stream import job_stream
+from timber.models.modeling_llama import LlamaForCausalLM, LlamaConfig
+from timber.utils import seed
+
 
 def load_vllm_model(args: ArgsType):
     from vllm import LLM
@@ -173,7 +166,7 @@ def load_model(args):
             except RuntimeError as e:
                 pass
 
-        model = model.to(infer_dtype)
+        # model = model.to(infer_dtype)
         print('lora checkpoint loaded from', args.checkpoint)
 
     elif args.method != 'none':
